+++ src/core.test.ts
@@ -11,6 +11,7 @@ jest.mock("fs", () => ({
 }));
 const writeFileSyncMock = writeFileSync as jest.Mock;
 const existsSyncMock = existsSync as jest.Mock;
+jest.spyOn(process, "cwd").mockReturnValue("");
 
 // eslint-disable-next-line @typescript-eslint/no-unsafe-return
 jest.mock("./utils.js", () => ({